+++ hospital_reservation/resources/views/hospital_menu/complete_download_rservation.blade.php
@@ -39,8 +39,8 @@ table {
 @endif
 
 @if($message = Session::get('success'))
-<div class ="alert alert-success alart-block">
-    <button type="button" class = "close" data-dismiss="alert">x</button>
+<div class ="alert alert-success alert-block">
+    <button type="button" class = "close" data-dismiss = "alert">x</button>
     <strong>{{ $message }}</strong>
 </div>
 @endif
